requirements.txt
aiohttp==3.6.2
amqp==2.5.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
idna==2.9
kaggle==1.5.6
kombu==4.6.8
multidict==4.7.5
python-dateutil==2.8.1
python-decouple==3.3
python-slugify==4.0.0
pytz==2019.3
ratelimit==2.2.1
requests==2.23.0
six==1.14.0
slackclient==2.5.0
sqlparse==0.3.1
text-unidecode==1.3
tqdm==4.44.1
urllib3==1.24.3
vine==1.3.0
yarl==1.4.2
